Added moving player
Created a new scene with a CharacterBody2D as the root node.
Added sprite as child and assigned the character image as the sprite.
Added collisionShape2D as child and created a new capsule shape and adjusted it to the sprite.
Added script to the root node and used 2d movement template